# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Smriti API
  description: Smarter Home for all your Photos and Videos
  version: 24.06.24
servers:
  - url: https://localhost:5001
...
paths:
...
  /v1/mediaItems:
    get:
      tags:
        - MediaItems
      parameters:
        - name: page
          in: query
          description: Pagination page
          schema:
            type: number
            default: 1
        - name: limit
          in: query
          description: Pagination limit
          schema:
            type: number
            default: 15
        - name: type
          in: query
          description: MediaItem Type
          schema:
            type: string
            example: video
        - name: category
          in: query
          description: MediaItem Category
          schema:
            type: string
            example: live
        - name: status
          in: query
          description: MediaItem Status
          schema:
            type: string
            example: FAILED
      summary: Get all existing mediaItems
      description: Get all existing mediaItems
      operationId: getMediaItems
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaItems'
        '400':
          description: Bad Request
        '401':
          description: Unauthenticated
        '500':
          description: Internal Server Error
    post:
      tags:
        - MediaItems
      summary: Upload a new mediaItem
      description: Upload a new mediaItem
      operationId: uploadMediaItem
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
        '401':
          description: Unauthenticated
        '500':
          description: Internal Server Error
  /v1/mediaItems/{id}:
    get:
      tags:
        - MediaItems
      parameters:
        - name: id
          in: path
          description: ID of mediaItem
          required: true
          schema:
            type: string
            format: uuid
      summary: Get an existing mediaItem
      description: Get an existing mediaItem
      operationId: getMediaItem
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaItem'
        '400':
          description: Bad Request
        '401':
          description: Unauthenticated
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
    put:
      tags:
        - MediaItems
      parameters:
        - name: id
          in: path
          description: ID of mediaItem
          required: true
          schema:
            type: string
            format: uuid
      summary: Update an existing mediaItem
      description: Update an existing mediaItem
      operationId: updateMediaItem
      security:
        - bearerAuth: []
      requestBody:
        description: MediaItem request body
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaItemRequest'
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthenticated
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
    delete:
      tags:
        - MediaItems
      parameters:
        - name: id
          in: path
          description: ID of mediaItem
          required: true
          schema:
            type: string
            format: uuid
      summary: Delete an existing mediaItem
      description: Delete an existing mediaItem
      operationId: deleteMediaItem
      security:
        - bearerAuth: []
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthenticated
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /v1/mediaItems/{id}/places:
    get:
      tags:
        - MediaItems
      parameters:
        - name: id
          in: path
          description: ID of mediaItem
          required: true
          schema:
            type: string
            format: uuid
      summary: Get places from an existing mediaItem
      description: Get places from an existing mediaItem
      operationId: getMediaItemPlaces
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Places'
        '400':
          description: Bad Request
        '401':
          description: Unauthenticated
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /v1/mediaItems/{id}/things:
    get:
      tags:
        - MediaItems
      parameters:
        - name: id
          in: path
          description: ID of mediaItem
          required: true
          schema:
            type: string
            format: uuid
      summary: Get things from an existing mediaItem
      description: Get things from an existing mediaItem
      operationId: getMediaItemThings
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Things'
        '400':
          description: Bad Request
        '401':
          description: Unauthenticated
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /v1/mediaItems/{id}/people:
    get:
      tags:
        - MediaItems
      parameters:
        - name: id
          in: path
          description: ID of mediaItem
          required: true
          schema:
            type: string
            format: uuid
      summary: Get people from an existing mediaItem
      description: Get people from an existing mediaItem
      operationId: getMediaItemPeople
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/People'
        '400':
          description: Bad Request
        '401':
          description: Unauthenticated
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /v1/mediaItems/{id}/albums:
    get:
      tags:
        - MediaItems
      parameters:
        - name: id
          in: path
          description: ID of mediaItem
          required: true
          schema:
            type: string
            format: uuid
      summary: Get albums from an existing mediaItem
      description: Get albums from an existing mediaItem
      operationId: getMediaItemAlbums
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Albums'
        '400':
          description: Bad Request
        '401':
          description: Unauthenticated
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
  schemas:
...
    MediaItemRequest:
      type: object
      properties:
        description:
          type: string
          example: Pictures from Wedding of Jane and Doe
        favourite:
          type: boolean
          example: true
        hidden:
          type: boolean
          example: false
    MediaItemFace:
      type: object
      properties:
        thumbnail:
          readOnly: true
          type: string
          example: base-64-thumbnail-image
    MediaItem:
      type: object
      properties:
        id:
          readOnly: true
          type: string
          format: uuid
          example: 102dc522de-1669-48a4-8270-4950a0237599
        userId:
          readOnly: true
          type: string
          format: uuid
          example: 102dc522de-1669-48a4-8270-4950a0237599
        filename:
          readOnly: true
          type: string
          example: IMG_4942.jpg
        description:
          readOnly: true
          type: string
          example: Picture of a pizza from New York
        mimeType:
          readOnly: true
          type: string
          example: image/jpeg
        placeholder:
          readOnly: true
          type: string
          example: base-64-placeholder-image
        sourceUrl:
          readOnly: true
          type: string
          example: https://cdn.smriti/source
        previewUrl:
          readOnly: true
          type: string
          example: https://cdn.smriti/preview
        thumbnailUrl:
          readOnly: true
          type: string
          example: https://cdn.smriti/thumbnail
        favourite:
          type: boolean
          example: false
        shared:
          type: boolean
          example: true
        hidden:
          type: boolean
          example: false
        deleted:
          type: boolean
          example: false
        status:
          readOnly: true
          type: string
          enum: ["UNSPECIFIED", "PROCESSING", "READY", "FAILED"]
        mediaItemType:
          readOnly: true
          type: string
          enum: ["photo", "video"]
        mediaItemCategory:
          readOnly: true
          type: string
          enum: ["default", "screenshot", "panorama", "slow", "motion", "live"]
        width:
          readOnly: true
          type: integer
          example: 1080
        height:
          readOnly: true
          type: integer
          example: 720
        creationTime:
          readOnly: true
          type: string
          format: date-time
          example: "2022-09-14T16:58:28Z"
        latitude:
          readOnly: true
          type: number
          example: 19.384982
        longitude:
          readOnly: true
          type: number
          example: -70.638324
        cameraMake:
          readOnly: true
          type: string
          example: Apple
        cameraModel:
          readOnly: true
          type: string
          example: iPhone 12
        focalLength:
          readOnly: true
          type: string
          example: 5.49 mm
        apertureFNumber:
          readOnly: true
          type: string
          example: f/1.8
        isoEquivalent:
          readOnly: true
          type: string
          example: ISO758
        exposureTime:
          readOnly: true
          type: string
          example: 1/20
        fps:
          readOnly: true
          type: string
          example: 60
        createdAt:
          readOnly: true
          type: string
          format: date-time
          example: "2022-09-14T16:58:28Z"
        updatedAt:
          readOnly: true
          type: string
          format: date-time
          example: "2022-09-14T16:58:28Z"
...
    Album:
      type: object
      properties:
        id:
          readOnly: true
          type: string
          format: uuid
          example: 102dc522de-1669-48a4-8270-4950a0237599
        userId:
          readOnly: true
          type: string
          format: uuid
          example: 102dc522de-1669-48a4-8270-4950a0237599
        name:
          type: string
          example: Jane's Wedding
        description:
          type: string
          example: Pictures from Wedding of Jane and Doe
        coverMediaItemId:
          type: string
          format: uuid
          example: 102dc522de-1669-48a4-8270-4950a0237599
        shared:
          type: boolean
          example: true
        hidden:
          type: boolean
          example: false
        mediaItemsCount:
          type: integer
          example: 42
        createdAt:
          readOnly: true
          type: string
          format: date-time
          example: "2022-09-14T16:58:28Z"
        updatedAt:
          readOnly: true
          type: string
          format: date-time
          example: "2022-09-14T16:58:28Z"
        coverMediaItem:
          $ref: '#/components/schemas/MediaItem'
    Place:
      type: object
      properties:
        id:
          readOnly: true
          type: string
          format: uuid
          example: 102dc522de-1669-48a4-8270-4950a0237599
        postcode:
          type: string
          example: 421201
        town:
          type: string
          example: Dallas
        city:
          type: string
          example: Mumbai
        state:
          type: string
          example: Washington
        country:
          type: string
          example: India
        coverMediaItemId:
          type: string
          format: uuid
          example: 102dc522de-1669-48a4-8270-4950a0237599
        createdAt:
          type: string
          format: date-time
          example: "2022-09-14T16:58:28Z"
        updatedAt:
          type: string
          format: date-time
          example: "2022-09-14T16:58:28Z"
        coverMediaItem:
          $ref: '#/components/schemas/MediaItem'
    Thing:
      type: object
      properties:
        id:
          readOnly: true
          type: string
          format: uuid
          example: 102dc522de-1669-48a4-8270-4950a0237599
        name:
          type: string
          example: Cars
        coverMediaItemId:
          type: string
          format: uuid
          example: 102dc522de-1669-48a4-8270-4950a0237599
        createdAt:
          type: string
          format: date-time
          example: "2022-09-14T16:58:28Z"
        updatedAt:
          type: string
          format: date-time
          example: "2022-09-14T16:58:28Z"
        coverMediaItem:
          $ref: '#/components/schemas/MediaItem'
...
    Person:
      type: object
      properties:
        id:
          readOnly: true
          type: string
          format: uuid
          example: 102dc522de-1669-48a4-8270-4950a0237599
        name:
          type: string
          example: Cars
        coverMediaItemId:
          type: string
          format: uuid
          example: 102dc522de-1669-48a4-8270-4950a0237599
        coverMediaItemFaceId:
          type: string
          format: uuid
          example: 102dc522de-1669-48a4-8270-4950a0237599
        createdAt:
          type: string
          format: date-time
          example: "2022-09-14T16:58:28Z"
        updatedAt:
          type: string
          format: date-time
          example: "2022-09-14T16:58:28Z"
        coverMediaItemFace:
          $ref: '#/components/schemas/MediaItemFace'
...
    MediaItems:
      type: array
      items:
        $ref: '#/components/schemas/MediaItem'
    Places:
      type: array
      items:
        $ref: '#/components/schemas/Place'
    Things:
      type: array
      items:
        $ref: '#/components/schemas/Thing'
    People:
      type: array
      items:
        $ref: '#/components/schemas/Person'
    Albums:
      type: array
      items:
        $ref: '#/components/schemas/Album'
...